+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added a test for typescript typings
+
 ## 1.2.0
 
 * Added typescript definitions